Rename head to first - replace head with first in main Annexe pkg - replace head with first in last pkg
