+++ packages/first/src/index.ts
@@ -1,3 +1,3 @@
-export function head<T>(value: T[]): T | undefined {
+export function first<T>(value: T[]): T | undefined {
   return Array.isArray(value) && value.length ? value[0] : void 0;
 }
